refactor(frontend): extract match tracking tab
